Only add whitelisted babel-plugins from registry to pipeline
In the previous commit a babel-plugin registered on the registry is
added to the transformation pipeline. Since this doesn't work as
expected for the case where an addon depends on such a babel-plugin but
the app which uses the addon doesn't include the babel-plugin, we
currently only support a whitelist of specific plugins until this is
figured out and eventually work.

This commit adds `ember-cli-htmlbars-inline-precompile` to the
whitelisted plugins and adds a test that other registered plugins are
not added to the transformation pipeline.