Properly detect host ember-cli version.
When running in a linked environment, the standard `require.resovle`
technique to determine the currently running ember-cli version does not
work properly (because the linked addon's ember-cli dev dependency will
be found instead of the _actual_ running ember-cli from the project).

The most common symptom of this issue is that modules are "double
transpiled" (e.g. you have an AMD wrapped module _inside_ of another AMD
wrapper).