Update babel-plugin-ember-modules-api-polyfill to 2.3.0.
The primary change here is that imports are lazily accessed where used,
instead of creating a variable for the import eagerly.

In practice this doesn't matter very much to folks, but in some
circumstances a given import may or may not end up being present (e.g.
`Ember.Test.registerWaiter`). The prior system would throw an error in
production builds when `import { registerWaiter } from '@ember/test';`
was used (due to the mechanism the prior version of the polyfill
worked) even if that import was only used in an `if (DEBUG) {` style
guard.

This update fixes that issue.