Ensure external helpers are not used when modules are not transpiled.
When using something akin to the following somewhat common pattern:

```js
let babel = this.addons.find(addon => addon.name === 'ember-cli-babel');
let tree = babel.transpileTree(input, {
  'ember-cli-babel': {
    compileModules: false,
  }
});
```

The previous logic would still attempt to use external helpers even when
modules themselves are not being transpiled. This meant that the final
output would include untranspiled `import` statements.

This affects common addons such as:

* ember-fetch
* ember-service-worker
* ember-data
* @ember/test-helpers
* ember-angle-bracket-invocation-polyfill
* ember-modifier-manager-polyfill

---

The fix here is to detect that modules will not be compiled, and avoid
usage of the external helpers.

This also moves the validation that parent addon's do not specify
`includeExternalHelpers` in their own config to the shared
`getAddonConfig` method (and makes the logic in that method a bit easier
to grok).