[BUGFIX] Dedupe internal decorators plugins
babel-plugin-filter-imports currently adds the internal
`decorators` plugin so it can parse and remove decorators,
but this conflicts with the _actual_ decorators plugin. This
PR dedupes the plugins for now (and for previous users of
babel-plugin-filter-imports), we'll also release a new major
version of that plugin that doesn't add the internal plugin.