[BUGFIX] Ensure class properties/decorator plugins are added in correct order
This PR uses the updated babel-plugin-helpers library to actually modify
the user's plugins and add the decorator/class properties transforms
relative to each other, if one shouldn't exist. They are also
constrained to come _before_ the Typescript transforms, should they
exist. This allows TS to react to any of the output of the previous
transforms.