Update babel-plugin-ember-modules-api-polyfill to fix issues with invalid identifiers.
The babel plugin previously took an identifier and replaced it's `name` value
with one that included a string like `Ember.Foo`. This is completely
invalid (a string like `Ember.Foo` is actually a MemberExpression), but
we got away with it because for the most part Babel happily printed
**whatever** value you give it in that spot.

Unfortunately, this happy mistake has caused us quite a bit of pain due
to Babel introducing new plugins that run as part of `@babel/preset-env`
(e.g. to make sure that all identifiers would work properly on IE9), and
those plugins **properly** sanitizing invalid identifiers.

The plugin was updated to avoid this mistake, and properly rewrite the
identifiers with a member expression. It also added tests to confirm
things are working as intended.