Avoid repeatedly checking the `ember-cli` version
Since `ember-cli` is the process controlling the addon instantiation, we
only need to ensure that we are using the correct version once (not once
per ember-cli-babel instance).

Since `ember-cli-babel` is the single most commonly used addon, this
non-trivially reduces the amount of overhead during the early part of a
build.